move resampler into babycat module
